+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft CMS 3.x
 
+## Unreleased
+
+### Added
+- Added `craft\helpers\App::phpSizeToBytes()`.
+
 ## 3.6.0-RC2 - 2020-12-15
 
 ### Added